requirements.txt
pytest
mypy
pylint
smbus2
pimoroni-bme280
RPi.GPIO
Mock.GPIO
freezegun